\name{pt}
\alias{ProbabilitePT}
\alias{GuiPT}
\alias{powerExample}
\alias{.First.lib}
\alias{vraispoistweedie}
\alias{parpoistweedie} 
\title{ Some discrete exponential dispersion models:Poisson-Tweedie}
\description{ This package provides a GUI for the poistweedie package, it is
        developed as an Rcmdr plug in.}


\usage{
.First.lib(libname, pkgname)
ProbabilitePT()
GuiPT()
powerExample()
vraispoistweedie(para,y)
parpoistweedie(y)
}

\arguments{
\item{libname}{Rcmdr}
\item{pkgname}{Poisson-Tweedie Plugin}	
\item{para}{vector of parameter}
\item{y}{vector of random data}	
		
}

\author{Cactha David Pechel, Laure Pauline Fotso  and  Celestin C Kokonendji
Maintainer: Cactha David Pechel ( <davidpechel@yahoo.fr>)}


\references{
Dunn, Peter K and Smyth, Gordon K (To appear).
Series evaluation of Tweedie exponential dispersion model densities
\emph{Statistics and Computing}.

Dunn, Peter K and Smyth, Gordon K (2001).
Tweedie family densities: methods of evaluation.
\emph{Proceedings of the 16th International Workshop on Statistical Modelling},
Odense, Denmark, 2--6 July

Hougaard, P., Lee, M-L.T. and Whitmore, G.A. (1997). Analysis of
overdispersed count data by mixtures of Poisson variables and Poisson processes,
\emph{Biometrics}
\bold{53}, 1225--1238

Jorgensen, B. (1987).
Exponential dispersion models.
\emph{Journal of the Royal Statistical Society}, B,
\bold{49}, 127--162.

Kokonendji, C.C., Demeetrio, C.G.B. and Dossou-Gbete, S. (2004). Some
discrete exponential dispersion models: Poisson-Tweedie and Hinde-Demetrio
classes. SORT: Statistics and Operations Research Transactions 
\bold{28} (2), 201--214. 

}

			
			
			
\examples{

			
}
		